Modify validating Ids in controller, and get data after created and updated

+++ root/backend/app/controllers/course.controller.ts
@@ -3,6 +3,10 @@ import { ENUM_ERROR_CODE } from "../enums/enums";
 import { Result } from "../../libs/Result";
 import CourseService from "../services/course.service";
 import { CourseData, CourseSubjectData } from "../models/course-model";
+import { ProgrammeData } from "../models/programme-model";
+import ProgrammeService from "../services/programme.service";
+import { SubjectData } from "../models/subject-model";
+import SubjectService from "../services/subject.service";
 
 export default class CourseController {
   async getAllCourses(req: Request, res: Response) {
@@ -10,7 +14,7 @@ export default class CourseController {
     const pageSize: number = parseInt(req.query.pageSize as string) || 15;
     const query: string = req.query.query as string || "";
 
-    const response: Result<CourseData[]> = await CourseService.getCourses(query, pageSize, page);
+    const response: Result<CourseData[]> = await CourseService.getAllCourses(query, pageSize, page);
 
     if (response.isSuccess()) {
       return res.sendSuccess.ok(response.getData(), response.getMessage());
@@ -41,6 +45,12 @@ export default class CourseController {
     const courseName: string = req.body.courseName;
     const programmeId: number = req.body.programmeId;
 
+    const programmeResponse: Result<ProgrammeData> = await ProgrammeService.getProgrammeById(programmeId);
+
+    if (!programmeResponse.isSuccess()) {
+      return res.sendError.notFound(programmeResponse.getMessage());
+    }
+
     const response = await CourseService.createCourse(courseName, programmeId);
 
     if (response.isSuccess()) {
@@ -58,6 +68,13 @@ export default class CourseController {
     const courseName: string = req.body.courseName;
     const programmeId: number = req.body.programmeId;
 
+    const programmeResponse: Result<ProgrammeData> = await ProgrammeService.getProgrammeById(programmeId);
+    const courseResponse: Result<CourseData> = await CourseService.getCourseById(courseId);
+
+    if (!programmeResponse.isSuccess() || !courseResponse.isSuccess()) {
+      return res.sendError.notFound("Invalid programmeId, or courseId");
+    }
+
     const response = await CourseService.updateCourseById(courseId, courseName, programmeId);
 
     if (response.isSuccess()) {
@@ -73,6 +90,12 @@ export default class CourseController {
   async deleteCourseById(req: Request, res: Response) {
     const courseId: number = parseInt(req.params.courseId as string);
 
+    const courseResponse: Result<CourseData> = await CourseService.getCourseById(courseId);
+
+    if (!courseResponse.isSuccess()) {
+      return res.sendError.notFound("Invalid courseId");
+    }
+    
     const response = await CourseService.deleteCourseById(courseId);
 
     if (response.isSuccess()) {
@@ -107,6 +130,20 @@ export default class CourseController {
     const courseId: number = req.body.courseId;
     const subjectId: number = req.body.subjectId;
 
+    const courseIdResponse: Result<CourseData> = await CourseService.getCourseById(courseId);
+
+    const subjectIdResponse: Result<SubjectData> = await SubjectService.getSubjectById(subjectId);
+
+    if (!courseIdResponse.isSuccess() || !subjectIdResponse.isSuccess()) {
+      return res.sendError.notFound("Invalid courseId or subjectId");
+    }
+
+    const isCourseSubjectExist: boolean = await CourseService.isCourseSubjectExist(courseId, subjectId);
+
+    if (isCourseSubjectExist) {
+      return res.sendError.notFound("Course subject already exist");
+    }
+
     const response = await CourseService.createCourseSubject(courseId, subjectId);
 
     if (response.isSuccess()) {
@@ -124,6 +161,20 @@ export default class CourseController {
   async deleteCourseSubjectByCourseIdAndSubjectId(req: Request, res: Response) {
     const courseId: number = parseInt(req.params.courseId as string);
     const subjectId: number = parseInt(req.params.subjectId as string);
+
+    const courseIdResponse: Result<CourseData> = await CourseService.getCourseById(courseId);
+
+    const subjectIdResponse: Result<SubjectData> = await SubjectService.getSubjectById(subjectId);
+
+    if (!courseIdResponse.isSuccess() || !subjectIdResponse.isSuccess()) {
+      return res.sendError.notFound("Invalid courseId or subjectId");
+    }
+
+    const isCourseSubjectExist: boolean = await CourseService.isCourseSubjectExist(courseId, subjectId);
+
+    if (!isCourseSubjectExist) {
+      return res.sendError.notFound("Course subject not found");
+    }
 
     const response = await CourseService.deleteCourseSubjectByCourseIdAndSubjectId(courseId, subjectId);
